Add handling of assigning tickets to inactive accounts.
